refactor: split main helpers into modules
