+++ src/main.ts
@@ -1,4 +1,4 @@
-import { createInitialState, resetState, spawnFleet, spawnShip, simulateStep } from './core/gameState.js';
+import { createInitialState, resetState, spawnFleet, spawnShip } from './core/gameState.js';
 // Import UI styles so webpack extracts them into a hashed CSS asset via MiniCssExtractPlugin
 import './styles/ui.css';
 // Ensure low-level GL/readPixels patches are applied as early as possible
@@ -19,7 +19,7 @@ import * as THREE from 'three';
 if (!(globalThis as any).THREE) {
   (globalThis as any).THREE = THREE;
 }
-import type { GameState, UIElements, ShipClass } from './types/index.js';
+import type { GameState, UIElements } from './types/index.js';
 import { createThreeRenderer } from './renderer/threeRenderer.js';
 import { shipInstancer } from './renderer/shipInstancer.js';
 import { RendererConfig } from './config/rendererConfig.js';
@@ -31,52 +31,20 @@ import { getSVGLoader, loadSVGAsset } from './core/svgLoader.js';
 import * as logger from './utils/logger.js';
 import { DefaultGameConfig } from './config/gameConfig.js';
 import { defaultSVGConfig, getShipSVGUrls } from './config/svgConfig.js';
-import { perf, perfBegin, perfEnd } from './utils/perf.js';
+import { perf } from './utils/perf.js';
+import { bindUI } from './ui/bindUI.js';
+import { randomClass } from './utils/randomShipClass.js';
+import { reFormFleets } from './core/reFormFleets.js';
+import { startLoops } from './core/startLoops.js';
+import { setupPerfOverlay } from './utils/perfOverlay.js';
+import { enablePerfCollectorIfRequested } from './utils/perfCollector.js';
+import { wireControls } from './ui/wireControls.js';
+import { setupCameraControls } from './renderer/cameraControls.js';
 
 // Strongly-typed runtime debug hooks exposed on globalThis when enabled.
 declare global {
   interface Window { __appDebug?: unknown; __shipInstancer?: typeof shipInstancer; DEBUG_SHIP_INSTANCER?: boolean; }
   interface GlobalEventHandlers { }
-}
-
-function $(id: string) { return document.getElementById(id)!; }
-
-function bindUI(): UIElements {
-  return {
-    canvas: document.getElementById('world') as HTMLCanvasElement,
-    startPause: $('startPause') as HTMLButtonElement,
-    reset: $('reset') as HTMLButtonElement,
-    addRed: $('addRed') as HTMLButtonElement,
-    addBlue: $('addBlue') as HTMLButtonElement,
-    toggleTrails: $('toggleTrails') as HTMLButtonElement,
-    speed: $('speed') as HTMLDivElement,
-    redScore: $('redScore') as HTMLDivElement,
-    blueScore: $('blueScore') as HTMLDivElement,
-    stats: $('stats') as HTMLDivElement,
-    continuous: $('continuousCheckbox') as HTMLInputElement,
-    seedBtn: $('seedBtn') as HTMLButtonElement,
-    formationBtn: $('formationBtn') as HTMLButtonElement,
-  };
-}
-
-function randomClass(state: GameState): ShipClass { return (['fighter','corvette','frigate','destroyer','carrier'] as const as ShipClass[])[Math.floor(state.rng.next()*5)]; }
-
-function reFormFleets(state: GameState) {
-  const leftX = FleetConfig.positioning.leftMargin;
-  const rightX = state.simConfig.simBounds.width - FleetConfig.positioning.rightMargin;
-  const centerZ = state.simConfig.simBounds.depth / 2 + FleetConfig.positioning.centerZOffset;
-  let ri = 0, bi = 0;
-  const row = FleetConfig.positioning.rowSize;
-  const spacing = FleetConfig.positioning.spacing;
-  for (const s of state.ships) {
-    const i = (s.team === 'red' ? ri++ : bi++);
-    const col = i % row; const r = Math.floor(i / row);
-    const x = s.team === 'red' ? leftX + col * spacing : rightX - col * spacing;
-    const y = 400 + r * spacing;
-    const z = centerZ + (state.rng.next() - 0.5) * 100; // Random z position around center
-    s.pos.x = x; s.pos.y = y; s.pos.z = z;
-    s.vel.x = 0; s.vel.y = 0; s.vel.z = 0;
-  }
 }
 
 function initGame(seed?: string) {
@@ -378,550 +346,16 @@ function initGame(seed?: string) {
   })();
 }
 
-function wireControls(state: GameState, ui: UIElements) {
-  function updateSpeedLabel() { ui.speed.textContent = `Speed: ${state.speedMultiplier}×`; }
-  function updateRunLabel() { ui.startPause.textContent = state.running ? '⏸ Pause' : '▶ Start'; }
-  function updateScores() { ui.redScore.textContent = `Red ${state.score.red}`; ui.blueScore.textContent = `Blue ${state.score.blue}`; }
 
-  ui.startPause.onclick = () => { state.running = !state.running; updateRunLabel(); };
-  ui.reset.onclick = () => { resetState(state); spawnFleet(state,'red', FleetConfig.spawning.defaultFleetSize); spawnFleet(state,'blue', FleetConfig.spawning.defaultFleetSize); reFormFleets(state); updateScores(); };
-  ui.addRed.onclick = () => { spawnShip(state, 'red', randomClass(state)); };
-  ui.addBlue.onclick = () => { spawnShip(state, 'blue', randomClass(state)); };
-  ui.toggleTrails.onclick = () => { RendererConfig.visual.enableTrails = !RendererConfig.visual.enableTrails; ui.toggleTrails.textContent = `☄ Trails: ${RendererConfig.visual.enableTrails ? 'On' : 'Off'}`; };
-  ui.speed.onclick = () => {
-    const seq = [0.5, 1, 2, 4] as const; const i = seq.indexOf(state.speedMultiplier as any);
-    const next = seq[(i + 1) % seq.length]; state.speedMultiplier = next; updateSpeedLabel();
-  };
-  ui.seedBtn.onclick = () => { const s = `SEED-${Date.now()}`; resetState(state, s); spawnFleet(state,'red', FleetConfig.spawning.defaultFleetSize); spawnFleet(state,'blue', FleetConfig.spawning.defaultFleetSize); reFormFleets(state); updateScores(); };
-  ui.formationBtn.onclick = () => {
-    // Clear existing ships
-    state.ships = [];
-    state.bullets = [];
-    // Spawn randomized fleets
-    const fleetSize = FleetConfig.spawning.defaultFleetSize;
-    for (let i = 0; i < fleetSize; i++) {
-      const redClass = randomClass(state);
-      const blueClass = randomClass(state);
-      spawnShip(state, 'red', redClass);
-      spawnShip(state, 'blue', blueClass);
-    }
-    updateScores();
-  };
 
-  updateSpeedLabel(); updateRunLabel(); updateScores();
-}
-
-function resetToCinematicView(state: GameState) {
-  if (!state.renderer || state.ships.length === 0) return;
-
-  // Calculate center of mass of all alive ships
-  let centerX = 0, centerY = 0, centerZ = 0;
-  let shipCount = 0;
-
-  for (const ship of state.ships) {
-    if (ship.health > 0) { // Only count alive ships
-      centerX += ship.pos.x;
-      centerY += ship.pos.y;
-      centerZ += ship.pos.z;
-      shipCount++;
-    }
-  }
-
-  if (shipCount === 0) return;
-
-  centerX /= shipCount;
-  centerY /= shipCount;
-  centerZ /= shipCount;
-
-  // Calculate the spread (bounding box) of ships to determine optimal zoom
-  let minX = Infinity, maxX = -Infinity;
-  let minY = Infinity, maxY = -Infinity;
-  let minZ = Infinity, maxZ = -Infinity;
-
-  for (const ship of state.ships) {
-    if (ship.health > 0) {
-      minX = Math.min(minX, ship.pos.x);
-      maxX = Math.max(maxX, ship.pos.x);
-      minY = Math.min(minY, ship.pos.y);
-      maxY = Math.max(maxY, ship.pos.y);
-      minZ = Math.min(minZ, ship.pos.z);
-      maxZ = Math.max(maxZ, ship.pos.z);
-    }
-  }
-
-  // Calculate dimensions of the ship cluster
-  const spreadX = maxX - minX;
-  const spreadY = maxY - minY;
-  const spreadZ = maxZ - minZ;
-  const maxSpread = Math.max(spreadX, spreadY, spreadZ);
-
-  // Set camera target to center of mass
-  state.renderer.cameraTarget.x = centerX;
-  state.renderer.cameraTarget.y = centerY;
-  state.renderer.cameraTarget.z = centerZ;
-
-  // Calculate optimal distance based on spread and camera FOV
-  const _fovRadians = (RendererConfig.camera.fov * Math.PI) / 180;
-  const optimalDistance = (maxSpread / 2) / Math.tan(_fovRadians / 2) * CameraConfig.resetToCinematic.fovMultiplier; // 1.5x for comfortable viewing
-
-  // Set distance with some minimum/maximum bounds
-  state.renderer.cameraDistance = Math.max(CameraConfig.cinematic.minDistance, Math.min(CameraConfig.cinematic.maxDistance, optimalDistance));
-
-  // Reset camera rotation to a good default viewing angle
-  state.renderer.cameraRotation.x = CameraConfig.resetToCinematic.cameraRotation.x; // Slight downward tilt
-  state.renderer.cameraRotation.y = CameraConfig.resetToCinematic.cameraRotation.y; // Face the action
-  state.renderer.cameraRotation.z = CameraConfig.resetToCinematic.cameraRotation.z;
-}
-
-function updateCinematicCamera(state: GameState, dt: number) {
-  if (!state.renderer || state.ships.length === 0) return;
-
-  // Separate ships by team
-  const redShips = state.ships.filter(s => s.team === 'red' && s.health > 0);
-  const blueShips = state.ships.filter(s => s.team === 'blue' && s.health > 0);
-
-  if (redShips.length === 0 || blueShips.length === 0) return;
-
-  // Calculate center of each fleet
-  let redCenterX = 0, redCenterY = 0, redCenterZ = 0;
-  let blueCenterX = 0, blueCenterY = 0, blueCenterZ = 0;
-
-  for (const ship of redShips) {
-    redCenterX += ship.pos.x;
-    redCenterY += ship.pos.y;
-    redCenterZ += ship.pos.z;
-  }
-  redCenterX /= redShips.length;
-  redCenterY /= redShips.length;
-  redCenterZ /= redShips.length;
-
-  for (const ship of blueShips) {
-    blueCenterX += ship.pos.x;
-    blueCenterY += ship.pos.y;
-    blueCenterZ += ship.pos.z;
-  }
-  blueCenterX /= blueShips.length;
-  blueCenterY /= blueShips.length;
-  blueCenterZ /= blueShips.length;
-
-  // Calculate the midpoint between both fleets
-  const centerX = (redCenterX + blueCenterX) / 2;
-  const centerY = (redCenterY + blueCenterY) / 2;
-  const centerZ = (redCenterZ + blueCenterZ) / 2;
-
-  // Calculate distance between fleets for optimal zoom
-  const dxF = redCenterX - blueCenterX;
-  const dyF = redCenterY - blueCenterY;
-  const dzF = redCenterZ - blueCenterZ;
-  const fleetDistance = Math.sqrt(dxF * dxF + dyF * dyF + dzF * dzF);
-
-  // Calculate optimal camera distance (show both fleets with some margin)
-  const _fovRadians = (RendererConfig.camera.fov * Math.PI) / 180;
-  const optimalDistance = Math.max(fleetDistance * CameraConfig.cinematic.fleetDistanceMultiplier, CameraConfig.cinematic.minDistance); // Minimum distance of 500
-
-  // Smoothly interpolate camera target
-  const lerpFactor = Math.min(dt * CameraConfig.cinematic.lerpFactor, 1); // Smooth following with 2 seconds lerp time
-  state.renderer.cameraTarget.x += (centerX - state.renderer.cameraTarget.x) * lerpFactor;
-  state.renderer.cameraTarget.y += (centerY - state.renderer.cameraTarget.y) * lerpFactor;
-  state.renderer.cameraTarget.z += (centerZ - state.renderer.cameraTarget.z) * lerpFactor;
-
-  // Smoothly interpolate camera distance
-  const distanceLerpFactor = Math.min(dt * CameraConfig.cinematic.distanceLerpFactor, 1); // Slower distance adjustment
-  state.renderer.cameraDistance += (optimalDistance - state.renderer.cameraDistance) * distanceLerpFactor;
-
-  // Clamp distance
-  state.renderer.cameraDistance = Math.max(CameraConfig.cinematic.minDistance, Math.min(CameraConfig.cinematic.maxDistance, state.renderer.cameraDistance));
-}
-
-function setupCameraControls(state: GameState, canvas: HTMLCanvasElement) {
-  if (!state.renderer) return;
-
-  let isMouseDown = false;
-  let lastMouseX = 0;
-  let lastMouseY = 0;
-  let isCinematicMode = false;
-
-  // Mouse controls for rotation (click and drag)
-  canvas.addEventListener('mousedown', (e) => {
-    isMouseDown = true;
-    lastMouseX = e.clientX;
-    lastMouseY = e.clientY;
-  });
-
-  canvas.addEventListener('mouseup', () => {
-    isMouseDown = false;
-  });
-
-  canvas.addEventListener('mousemove', (e) => {
-    if (!isMouseDown || !state.renderer) return;
-
-    const sensitivity = CameraConfig.controls.mouseSensitivity;
-    const deltaX = (e.clientX - lastMouseX) * sensitivity;
-    const deltaY = (e.clientY - lastMouseY) * sensitivity;
-
-    // Update camera rotation
-    // Orbit controls: yaw left/right with horizontal drag; pitch up/down with vertical drag
-    // Note: invert Y so dragging up rotates camera up (natural control)
-    state.renderer.cameraRotation.y -= deltaX;
-    state.renderer.cameraRotation.x -= deltaY;
-
-    // Clamp vertical rotation to prevent flipping
-  const pitchLimit = (Math.PI / 2) - 0.001;
-  state.renderer.cameraRotation.x = Math.max(-pitchLimit, Math.min(pitchLimit, state.renderer.cameraRotation.x));
-  // Keep yaw within [-PI, PI] to avoid numerical drift
-  if (state.renderer.cameraRotation.y > Math.PI) state.renderer.cameraRotation.y -= Math.PI * 2;
-  if (state.renderer.cameraRotation.y < -Math.PI) state.renderer.cameraRotation.y += Math.PI * 2;
-
-    lastMouseX = e.clientX;
-    lastMouseY = e.clientY;
-  });
-
-  // Mouse wheel for zoom (move camera forward/back)
-  canvas.addEventListener('wheel', (e) => {
-    if (!state.renderer) return;
-    e.preventDefault();
-
-    const zoomSpeed = CameraConfig.controls.zoomSpeed; // Fixed zoom speed
-  const zoomDirection = e.deltaY > 0 ? 1 : -1; // Positive deltaY = zoom out (move back), negative = zoom in (move forward)
-
-    // Calculate camera's forward vector
-    const pitch = state.renderer.cameraRotation.x;
-    const yaw = state.renderer.cameraRotation.y;
-    const forwardX = Math.cos(yaw) * Math.cos(pitch);
-    const forwardY = Math.sin(pitch);
-    const forwardZ = Math.sin(yaw) * Math.cos(pitch);
-
-  // Move camera and target together along forward vector
-  const moveDistance = zoomSpeed * zoomDirection;
-  // Use camera forward as direction; positive zoomDirection moves outward
-  state.renderer.cameraTarget.x += forwardX * moveDistance;
-  state.renderer.cameraTarget.y += forwardY * moveDistance;
-  state.renderer.cameraTarget.z += forwardZ * moveDistance;
-  });
-
-  // Keyboard controls for movement
-  const keys: { [key: string]: boolean } = {};
-  document.addEventListener('keydown', (e) => {
-    keys[e.code] = true;
-
-    // Prevent browser defaults on movement keys (e.g., Space scrolling)
-    if (e.code === 'KeyW' || e.code === 'KeyA' || e.code === 'KeyS' || e.code === 'KeyD' || e.code === 'ShiftLeft' || e.code === 'Space') {
-      e.preventDefault();
-    }
-
-    // Cinematic camera with 'C' key
-    if (e.code === 'KeyC' && state.renderer) {
-      e.preventDefault();
-      isCinematicMode = true;
-      resetToCinematicView(state); // Initial setup
-      logger.info('🎬 Cinematic camera mode activated - following both fleets');
-    }
-  });
-
-  document.addEventListener('keyup', (e) => {
-    keys[e.code] = false;
-  });
-
-  // Update camera position based on WASD keys
-  function updateCameraMovement(dt: number) {
-    if (!state.renderer) return;
-
-    const moveSpeed = CameraConfig.controls.moveSpeed * dt;
-    let moveForward = 0, moveRight = 0, moveUp = 0;
-  // W should move the camera forward (toward where the camera is looking).
-  if (keys['KeyW']) moveForward += moveSpeed;   // Forward (towards forward vector)
-  if (keys['KeyS']) moveForward -= moveSpeed;   // Backward
-    if (keys['KeyD']) moveRight   += moveSpeed;   // Right
-    if (keys['KeyA']) moveRight   -= moveSpeed;   // Left
-    if (keys['Space']) moveUp     += moveSpeed;   // Up
-    if (keys['ShiftLeft']) moveUp -= moveSpeed;   // Down
-
-    // Check if any movement keys are pressed to exit cinematic mode
-    const hasMovementInput = keys['KeyW'] || keys['KeyS'] || keys['KeyA'] || keys['KeyD'] || keys['ShiftLeft'] || keys['Space'];
-    if (hasMovementInput && isCinematicMode) {
-      isCinematicMode = false;
-      logger.info('🎮 Manual camera control activated - cinematic mode disabled');
-    }
-
-    // If cinematic mode is active and no manual input, update cinematic camera
-    if (isCinematicMode && !hasMovementInput) {
-      updateCinematicCamera(state, dt);
-      return; // Skip manual movement when in cinematic mode
-    }
-
-    // Build an orthonormal basis from yaw/pitch to avoid warped motion
-    const pitch = state.renderer.cameraRotation.x;
-    const yaw = state.renderer.cameraRotation.y;
-    // Forward (where camera looks)
-    let fx = Math.cos(yaw) * Math.cos(pitch);
-    let fy = Math.sin(pitch);
-    let fz = Math.sin(yaw) * Math.cos(pitch);
-  // Compute squared length and do a single sqrt for normalization (avoid Math.hypot allocation)
-  const flSq = fx * fx + fy * fy + fz * fz;
-  const fl = Math.sqrt(flSq) || 1;
-  fx /= fl; fy /= fl; fz /= fl;
-  // The spherical coords above describe the camera position offset from the target.
-  // The camera's forward vector (where it looks) points from the camera into the scene -
-  // i.e., from camera position toward the target, which is the negative of the offset.
-  // Negate to align with THREE.Camera.getWorldDirection and other codepaths.
-  fx = -fx; fy = -fy; fz = -fz;
-
-    // Right = normalize(cross(forward, worldUp))
-    const upWorldX = 0, upWorldY = 1, upWorldZ = 0;
-    let rx = fy * upWorldZ - fz * upWorldY; // fy*0 - fz*1 = -fz
-    let ry = fz * upWorldX - fx * upWorldZ; // fz*0 - fx*0 = 0
-    let rz = fx * upWorldY - fy * upWorldX; // fx*1 - fy*0 = fx
-  const rlSq = rx * rx + ry * ry + rz * rz;
-  const rl = Math.sqrt(rlSq) || 1;
-  rx /= rl; ry /= rl; rz /= rl;
-
-    // Up (camera-up) = normalize(cross(right, forward))
-    let ux = ry * fz - rz * fy;
-    let uy = rz * fx - rx * fz;
-    let uz = rx * fy - ry * fx;
-  const ulSq = ux * ux + uy * uy + uz * uz;
-  const ul = Math.sqrt(ulSq) || 1;
-  ux /= ul; uy /= ul; uz /= ul;
-
-  // Compose movement in world space: forward is along camera forward vector
-  const worldMoveX = moveRight * rx + moveUp * upWorldX + moveForward * fx;
-  const worldMoveY = moveRight * ry + moveUp * upWorldY + moveForward * fy;
-  const worldMoveZ = moveRight * rz + moveUp * upWorldZ + moveForward * fz;
-
-    state.renderer.cameraTarget.x += worldMoveX;
-    state.renderer.cameraTarget.y += worldMoveY;
-    state.renderer.cameraTarget.z += worldMoveZ;
-  }
-
-  // Add camera movement to the render loop
-  const originalRender = state.renderer.render;
-  state.renderer.render = function(dt: number) {
-    updateCameraMovement(dt);
-    originalRender.call(state.renderer, dt);
-  };
-}
-
-function startLoops(state: GameState, ui: UIElements) {
-  const fixedDt = 1 / state.simConfig.tickRate;
-  let last = performance.now();
-  let acc = 0;
-  let fpsAccum = 0, fpsFrames = 0, _fpsTime = 0;
-
-  function frame(now: number) {
-    perfBegin('frame.total');
-    
-    const dt = (now - last) / 1000; last = now;
-    acc += dt;
-    
-    if (state.running) {
-      // Fixed-step simulation with speed multiplier
-      perfBegin('simulation.step');
-      const maxSteps = 5;
-      let steps = 0;
-      while (acc >= fixedDt && steps < maxSteps) {
-        simulateStep(state, fixedDt * state.speedMultiplier);
-        try { state.physicsStepper?.step(fixedDt * state.speedMultiplier); } catch (_e) { void _e;/* ignore if missing */ }
-      state.time += fixedDt * state.speedMultiplier; state.tick++; state.frame = (state.frame ?? 0) + 1;
-        acc -= fixedDt; steps++;
-      }
-      perfEnd('simulation.step');
-      
-      // Auto-respawn if continuous
-      perfBegin('game.respawn');
-      if ((ui.continuous.checked)) {
-        const redAlive = state.ships.some(s => s.team === 'red');
-        const blueAlive = state.ships.some(s => s.team === 'blue');
-        if (!redAlive) spawnFleet(state, 'red', FleetConfig.spawning.defaultFleetSize);
-        if (!blueAlive) spawnFleet(state, 'blue', FleetConfig.spawning.defaultFleetSize);
-      }
-      perfEnd('game.respawn');
-    }
-
-    // Render
-    perfBegin('renderer.total');
-    state.renderer?.render(dt);
-    perfEnd('renderer.total');
-
-    // Stats
-    perfBegin('ui.stats');
-    fpsAccum += dt; fpsFrames++; _fpsTime += dt;
-    if (fpsAccum >= 0.5) {
-      const fps = Math.round(fpsFrames / fpsAccum);
-      ui.stats.textContent = `FPS ${fps} • Ships ${state.ships.length} • Bullets ${state.bullets.length} • Tick ${state.tick}`;
-      // Update scoreboard periodically
-      (document.getElementById('redScore') as HTMLDivElement).textContent = `Red ${state.score.red}`;
-      (document.getElementById('blueScore') as HTMLDivElement).textContent = `Blue ${state.score.blue}`;
-      fpsAccum = 0; fpsFrames = 0;
-    }
-    perfEnd('ui.stats');
-
-    perfEnd('frame.total');
-    requestAnimationFrame(frame);
-  }
-  requestAnimationFrame(frame);
-}
-
-function setupPerfOverlay() {
-  // Only show overlay if enabled via query param or config
-  const showOverlay = perf.isEnabled() && (
-    typeof location !== 'undefined' && location.search.includes('showPerf=1')
-  );
-  
-  if (!showOverlay) return;
-
-  // Create overlay element
-  const overlay = document.createElement('div');
-  overlay.id = 'perfOverlay';
-  overlay.style.cssText = `
-    position: fixed;
-    top: 10px;
-    right: 10px;
-    background: rgba(0, 0, 0, 0.8);
-    color: #00ff00;
-    padding: 10px;
-    font-family: 'Courier New', monospace;
-    font-size: 12px;
-    line-height: 1.3;
-    border-radius: 4px;
-    z-index: 1000;
-    min-width: 200px;
-    max-width: 300px;
-  `;
-
-  document.body.appendChild(overlay);
-
-  // Update overlay periodically
-  let lastUpdate = 0;
-  const updateInterval = 250; // 4fps
-
-  function updateOverlay() {
-    const now = performance.now();
-    if (now - lastUpdate < updateInterval) {
-      requestAnimationFrame(updateOverlay);
-      return;
-    }
-    lastUpdate = now;
-
-    const summary = perf.getSummary();
-    if (summary.frameCount === 0) {
-      overlay.textContent = 'Performance Monitor\nNo data yet...';
-      requestAnimationFrame(updateOverlay);
-      return;
-    }
-
-    // Build display content
-    let content = 'Performance Monitor\n';
-    content += `Frames: ${summary.frameCount}\n`;
-    content += `Total: ${summary.totalFrameMs}ms\n`;
-    content += `Avg: ${summary.avgFrameMs}ms\n`;
-    content += `P95: ${summary.p95FrameMs}ms\n\n`;
-
-    // Show top 8 subsystems
-    const topSubsystems = summary.subsystems.slice(0, 8);
-    for (const subsystem of topSubsystems) {
-      const percentage = summary.totalFrameMs > 0 
-        ? Math.round((subsystem.totalMs / summary.totalFrameMs) * 100)
-        : 0;
-      
-      content += `${subsystem.name}: ${percentage}% (${subsystem.avgMs}ms)\n`;
-    }
-
-    overlay.textContent = content;
-    requestAnimationFrame(updateOverlay);
-  }
-
-  requestAnimationFrame(updateOverlay);
-}
 
 // Boot
 // Initialize logger debug mode from global game config so devs can toggle verbosity
 logger.setDebug(!!DefaultGameConfig.ui.showDebugInfo);
 
 // Helper: enable perf collector when URL has ?debugPerf=1. Safe to call multiple times.
-function enablePerfCollectorIfRequested() {
-  try {
-    const enabled = typeof location !== 'undefined' && location.search.includes('debugPerf=1');
-    if (!enabled) {
-      return;
-    }
-    if ((window as any).__perf && typeof (window as any).__perf.addEvent === 'function') {
-      try { console.debug('[perf] collector already active (init)'); } catch {}
-      return;
-    }
-    const perf = {
-      _frameTimes: [] as number[],
-      _events: [] as { name: string; ms: number; t: number }[],
-      addEvent(e: { name: string; ms: number }) {
-        this._events.push({ name: e.name, ms: e.ms, t: performance.now() });
-        if (this._events.length > 2000) this._events.shift();
-      },
-      startFpsSampling() {
-        let last = performance.now();
-        const loop = (ts: number) => {
-          const dt = ts - last; last = ts;
-          this._frameTimes.push(dt);
-          if (this._frameTimes.length > 600) this._frameTimes.shift();
-          requestAnimationFrame(loop);
-        };
-        requestAnimationFrame(loop);
-      },
-      getFpsStats() {
-        const arr = this._frameTimes.slice();
-        if (!arr.length) return { avgFps: 0, p99FrameMs: 0 };
-        const avg = arr.reduce((a: number, b: number)=>a+b,0) / arr.length;
-        const fps = 1000 / avg;
-        const sorted = arr.slice().sort((a: number, b: number)=>a-b);
-        const idx = Math.max(0, Math.floor(sorted.length * 0.99) - 1);
-        return { avgFps: fps, p99FrameMs: sorted[idx] };
-      },
-      getEvents() { return this._events.slice(); }
-    } as any;
-    (window as any).__perf = perf;
-    perf.startFpsSampling();
-    try { console.debug('[perf] collector enabled (init)'); } catch {}
-  } catch { /* ignore */ }
-}
-
 window.addEventListener('DOMContentLoaded', () => { 
   try { enablePerfCollectorIfRequested(); } catch {}
   initGame();
 });
 
-// Lightweight main-thread perf collector, enabled via ?debugPerf=1
-(() => {
-  try {
-    const enabled = typeof location !== 'undefined' && location.search.includes('debugPerf=1');
-    if (!enabled) return;
-    const perf = {
-      _frameTimes: [] as number[],
-      _events: [] as { name: string; ms: number; t: number }[],
-      addEvent(e: { name: string; ms: number }) {
-        this._events.push({ name: e.name, ms: e.ms, t: performance.now() });
-        if (this._events.length > 2000) this._events.shift();
-      },
-      startFpsSampling() {
-        let last = performance.now();
-        const loop = (ts: number) => {
-          const dt = ts - last; last = ts;
-          this._frameTimes.push(dt);
-          if (this._frameTimes.length > 600) this._frameTimes.shift();
-          requestAnimationFrame(loop);
-        };
-        requestAnimationFrame(loop);
-      },
-      getFpsStats() {
-        const arr = this._frameTimes.slice();
-        if (!arr.length) return { avgFps: 0, p99FrameMs: 0 };
-        const avg = arr.reduce((a: number, b: number)=>a+b,0) / arr.length;
-        const fps = 1000 / avg;
-        const sorted = arr.slice().sort((a: number, b: number)=>a-b);
-        const idx = Math.max(0, Math.floor(sorted.length * 0.99) - 1);
-        return { avgFps: fps, p99FrameMs: sorted[idx] };
-      },
-      getEvents() { return this._events.slice(); }
-    } as any;
-    (window as any).__perf = perf;
-    perf.startFpsSampling();
-    try { console.debug('[perf] collector enabled'); } catch {}
-  } catch { /* ignore */ }
-})();
