+++ src/ui/bindUI.ts
@@ -1,0 +1,23 @@
+import type { UIElements } from '../types/index.js';
+
+function $(id: string) {
+  return document.getElementById(id)!;
+}
+
+export function bindUI(): UIElements {
+  return {
+    canvas: document.getElementById('world') as HTMLCanvasElement,
+    startPause: $('startPause') as HTMLButtonElement,
+    reset: $('reset') as HTMLButtonElement,
+    addRed: $('addRed') as HTMLButtonElement,
+    addBlue: $('addBlue') as HTMLButtonElement,
+    toggleTrails: $('toggleTrails') as HTMLButtonElement,
+    speed: $('speed') as HTMLDivElement,
+    redScore: $('redScore') as HTMLDivElement,
+    blueScore: $('blueScore') as HTMLDivElement,
+    stats: $('stats') as HTMLDivElement,
+    continuous: $('continuousCheckbox') as HTMLInputElement,
+    seedBtn: $('seedBtn') as HTMLButtonElement,
+    formationBtn: $('formationBtn') as HTMLButtonElement,
+  };
+}
